[misc] reporting 1.1.3, 1.1.4, 1.2.0 and 1.2.1 correction keeping R2DBC spec 0.9.1 compatibility

+++ src/main/java/org/mariadb/r2dbc/api/MariadbResult.java
@@ -13,7 +13,7 @@ import reactor.core.publisher.Mono;
 public interface MariadbResult extends Result {
 
   @Override
-  Mono<Long> getRowsUpdated();
+  Mono<Integer> getRowsUpdated();
 
   @Override
   <T> Flux<T> map(BiFunction<Row, RowMetadata, ? extends T> mappingFunction);
